+++ modules/meshcloud-mca-service-principal/module.tf
@@ -7,11 +7,11 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">=3.81.0"
+      version = ">=4.11.0"
     }
     azuread = {
       source  = "hashicorp/azuread"
-      version = ">=2.46.0"
+      version = ">=3.0.2"
     }
     azapi = {
       source  = "Azure/azapi"